chore: refactor triples-for by splitting it into multiple files
