+++ src/constants.js
@@ -1,0 +1,1 @@
+export const URI_TEMPLATE = "http://data.lblod.info/form-data/nodes/";
