+++ src/index.js
@@ -7,6 +7,8 @@ export * from "./import-triples-for-form.js";
 export * from "./namespaces.js";
 export * from "./remove-simple-form-value.js";
 export * from "./remove-triples.js";
-export * from "./triples-for.js";
+export * from "./triples-for/triples-for-path.js";
+export * from "./triples-for/triples-for-scope.js";
+export * from "./constants.js";
 export * from "./update-simple-form-value.js";
 export * from "./validation.js";
